requirements.txt
docker==6.1.3
click==8.1.7
rich==13.7.0
trivy-sdk==0.5.5
PyYAML==6.0.1
requests==2.31.0
python-debian==0.1.49
Flask==3.0.2
Flask-WTF==1.2.1
Bootstrap-Flask==2.3.3 